requirements.txt
asttokens==2.1.0
backcall==0.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
idna==3.4
ipykernel==6.17.1
ipython==8.6.0
jedi==0.18.1
jupyter_client==7.4.5
jupyter_core==5.0.0
kafka-python==2.0.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.4
prompt-toolkit==3.0.32
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.5
Pygments==2.13.0
pyparsing==3.0.9
pyspark==3.3.1
python-dateutil==2.8.2
pyzmq==24.0.1
requests==2.28.1
six==1.16.0
stack-data==0.6.1
tornado==6.2
traitlets==5.5.0
urllib3==1.26.12
wcwidth==0.2.5
